import more kinds of sprites

+++ src/pixel_buffer.rs
@@ -53,7 +53,7 @@ impl PixelBuffer {
                 let target_pixel = target_pixel as usize * BYTES_PER_PIXEL;
 
                 let color_index = sprite[x + (y * sprite_width)];
-                let color = palette[color_index.0];
+                let color = palette[color_index.0 as usize];
                 self.pixel_buffer[target_pixel..target_pixel + BYTES_PER_PIXEL]
                     .copy_from_slice(&color);
             });
